// src/api/vendor/controllers/requests/UpdateVendorDocument.ts
/*
 * spurtcommerce API
 * version 5.2.0
 * Copyright (c) 2021 piccosoft ltd
 * Author piccosoft ltd <support@piccosoft.com>
 * Licensed under the MIT license.
 */

import 'reflect-metadata';
import { IsNotEmpty } from 'class-validator';

export class UpdateVendorDocumentRequest {

    @IsNotEmpty()
    public certificate: Certificate;

}
interface Certificate {
    certificationType: string;
    refrenceNo: string;
    name: string;
    issuedBy: string;
    validFrom: string;
    validTo: string;
}